Added vagrant file and bootstrap.
The Vagrantfile and bootstrap.sh is used to start up a VM (using virtualbox)
and provision (or setup) all of the dependencies.  The command `vagrant
up` will cause vagrant to download the specified box (ubuntu), update
and install any packages located in bootstrap.sh and run finally run
commands located in bootstrap.  The ending command in our bootstrap run
the flask server on port 5000 and daemonizes the process.  Finally port
5000 is forwarded to port 50000 on the localhost.