A working version of our Scrum tasks.
Includes the ability to log in as students or faculty.
Faculty have the option to upload a variety of text, doc, or pdf files
to the webserver.  Faculty also have a Student view section to see how
uploading files appears for students.  Students can log in, if teacher
with the same class_id uploads documents, the documents will be
available to the student to download or view.